perf(bindings): split heavy export TUs by tracker family to reduce per-TU peak RAM
Split endgame, zero_dim, nid, and tracker export files into per-tracker-family
translation units (AMP, fixed-double, fixed-MP) using the coordinator pattern:
coordinator .cpp files use forward declarations only (no heavy headers), while
per-family files each include the heavy headers but instantiate only 1/3 of the
original template combinations.

Template visitor bodies moved from .cpp to .hpp (correct placement — template
member function definitions must be visible in every TU that instantiates them).

Before: zero_dim_export.cpp.o = 154 MB (6 instantiations), endgame = 57 MB,
nid = 51 MB. After: largest single TU is zero_dim_amp_export.cpp.o ≈ 119 MB,
with double/mp variants ≈ 101–109 MB — enabling 4-parallel builds without OOM.

New files: endgame_{config,double,mp,amp}_export.cpp,
tracker_config_export.cpp, zero_dim_{configs,double,mp,amp}_export.cpp,
nid_{datatypes,double,mp,amp}_export.cpp (13 files total).
