+++ python_bindings/src/endgame_amp_export.cpp
@@ -1,0 +1,33 @@
+// python/endgame_amp_export.cpp — AMPTracker endgame registrations.
+
+#include "endgame_export.hpp"
+
+namespace bertini{
+	namespace python{
+
+		void ExportAMPPSEG()
+		{
+			using TrackerT = AMPTracker;
+			using EGT = typename endgame::EndgameSelector<TrackerT>::PSEG;
+
+			class_<EGT>("AMPPSEG",
+				"The adaptive precision implementation of the power series endgame.",
+				init<TrackerT const&>( (arg("self"),arg("tracker")), "Default construct with default settings") )
+			.def(EndgameBaseVisitor<EGT>())
+			.def(PowerSeriesVisitor<EGT>());
+		}
+
+		void ExportAMPCauchyEG()
+		{
+			using TrackerT = AMPTracker;
+			using EGT = typename endgame::EndgameSelector<TrackerT>::Cauchy;
+
+			class_<EGT>("AMPCauchyEG",
+				"The adaptive precision implementation of the Cauchy endgame",
+				init<TrackerT const&>( (arg("self"),arg("tracker")), "Default construct with default settings"))
+			.def(EndgameBaseVisitor<EGT>())
+			.def(CauchyVisitor<EGT>())
+			;
+		}
+
+}} // namespaces
