+++ python_bindings/src/zero_dim_amp_export.cpp
@@ -1,0 +1,14 @@
+// python/zero_dim_amp_export.cpp — AMPTracker ZeroDim registrations.
+
+#include "zero_dim_export.hpp"
+
+namespace bertini{
+	namespace python{
+
+		void ExportZDAMP(){
+			using TrackerT = bertini::tracking::AMPTracker;
+			ExportZeroDimSpecific<TrackerT, bertini::endgame::EndgameSelector<TrackerT>::PSEG, bertini::System, bertini::start_system::TotalDegree>("ZeroDimPowerSeriesAdaptivePrecisionTotalDegree");
+			ExportZeroDimSpecific<TrackerT, bertini::endgame::EndgameSelector<TrackerT>::Cauchy, bertini::System, bertini::start_system::TotalDegree>("ZeroDimCauchyAdaptivePrecisionTotalDegree");
+		}
+
+}} // namespaces
